vm: move IntToBytes and BytesToInt to emit package
