+++ pkg/vm/emit/bigint.go
@@ -1,4 +1,4 @@
-package vm
+package emit
 
 import (
 	"encoding/binary"
