+++ pkg/vm/emit/bigint_test.go
@@ -1,4 +1,4 @@
-package vm
+package emit
 
 import (
 	"math"
